# Sort merge join: do not concatenate clusters