This PR is a rather simple change for the SMJ. Right now, after the actual merge is done (which is done on multiple clusters in parallel for large joins), the clusters are concatenated into a single large chunk at the end.

This has various problems: (i) it is unnecessary and slows the join down, (ii) it prohibits exploiting the sorted flag (given per chunk, but no longer possible when concatenating), and (iii) is hinders the following operators from parallelizing over chunks.

This PR no longer concatenates chunks and further parallelizes the output writing as it has shown to be significant in some situations (does not matter for ST execution).