@Bouncner 2020-02-05
Closing for now as there is an issue with debug builds failing which run just fine in release.