@Bouncner 2020-02-06
Hash Join performance:
```
----------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations
----------------------------------------------------------------------------
BM_Join_SmallAndSmall<JoinHash>       255442 ns       255447 ns         2051
BM_Join_SmallAndBig<JoinHash>      159634924 ns    159602007 ns            5
BM_Join_MediumAndMedium<JoinHash>   15220566 ns     15220865 ns           44
```

Old SortMergeJoin:
```
---------------------------------------------------------------------------------
Benchmark                                       Time             CPU   Iterations
---------------------------------------------------------------------------------
BM_Join_SmallAndSmall<JoinSortMerge>       267951 ns       267956 ns         2629
BM_Join_SmallAndBig<JoinSortMerge>      620472431 ns    620473053 ns            1
BM_Join_MediumAndMedium<JoinSortMerge>   18120199 ns     18120300 ns           37
```

New SortMergeJoin:
```
---------------------------------------------------------------------------------
Benchmark                                       Time             CPU   Iterations
---------------------------------------------------------------------------------
BM_Join_SmallAndSmall<JoinSortMerge>       269358 ns       269350 ns         2576
BM_Join_SmallAndBig<JoinSortMerge>      593867779 ns    593871558 ns            1
BM_Join_MediumAndMedium<JoinSortMerge>   16502659 ns     16500681 ns           41
```

Benchmarks are noisy. There is a small improvement (especially for large joins), but not much. More important is the ability now to mark chunks as sorted.