@Bouncner 2020-02-06
In case somebody wonders: for very large joins the SMJ remains to be faster than the HashJoin (that is without any bloom filters, which can have a vast impact in realistic joins) but it moved to 2Mx2M joins. So nothing we will probably see soon.